Move Linux-specific resolvers in the linux directory
Move all files that defines Facter::Resolvers::Linux::* into the
lib/facter/resolvers/linux directory.  Also only load them on Linux
systems as they are Linux-specific.
