+++ lib/facter/framework/core/file_loader.rb
@@ -54,14 +54,10 @@ require_relative '../../resolvers/az'
 require_relative '../../resolvers/base_resolver'
 require_relative '../../resolvers/containers'
 require_relative '../../resolvers/debian_version'
-require_relative '../../resolvers/disks'
-require_relative '../../resolvers/dmi'
 require_relative '../../resolvers/dmi_decode'
 require_relative '../../resolvers/ec2'
 require_relative '../../resolvers/eos_release'
 require_relative '../../resolvers/facterversion'
-require_relative '../../resolvers/filesystems'
-require_relative '../../resolvers/fips_enabled'
 require_relative '../../resolvers/gce'
 require_relative '../../resolvers/hostname'
 require_relative '../../resolvers/identity'
@@ -69,15 +65,12 @@ require_relative '../../resolvers/load_averages'
 require_relative '../../resolvers/lpar'
 require_relative '../../resolvers/lsb_release'
 require_relative '../../resolvers/lspci'
-require_relative '../../resolvers/memory'
 require_relative '../../resolvers/mountpoints'
 require_relative '../../resolvers/networking'
 require_relative '../../resolvers/open_vz'
 require_relative '../../resolvers/os_release'
 require_relative '../../resolvers/partitions'
 require_relative '../../resolvers/path'
-require_relative '../../resolvers/processors'
-require_relative '../../resolvers/processors_lscpu'
 require_relative '../../resolvers/redhat_release'
 require_relative '../../resolvers/release_from_first_line'
 require_relative '../../resolvers/ruby'
@@ -510,10 +503,17 @@ os_hierarchy.each do |os|
     require_relative '../../facts/linux/virtual'
     require_relative '../../facts/linux/xen'
 
+    require_relative '../../resolvers/linux/disks'
+    require_relative '../../resolvers/linux/dmi'
     require_relative '../../resolvers/linux/docker_uptime'
+    require_relative '../../resolvers/linux/filesystems'
+    require_relative '../../resolvers/linux/fips_enabled'
     require_relative '../../resolvers/linux/hostname'
     require_relative '../../resolvers/linux/load_averages'
+    require_relative '../../resolvers/linux/memory'
     require_relative '../../resolvers/linux/networking'
+    require_relative '../../resolvers/linux/processors'
+    require_relative '../../resolvers/linux/processors_lscpu'
 
   when 'linuxmint'
     require_relative '../../facts/linuxmint/os/name'
