+++ lib/facter/resolvers/linux/disks.rb
