+++ lib/facter/resolvers/linux/dmi.rb
