+++ lib/facter/resolvers/linux/filesystems.rb
