+++ lib/facter/resolvers/linux/fips_enabled.rb
