+++ lib/facter/resolvers/linux/memory.rb
