+++ lib/facter/resolvers/linux/processors.rb
