+++ lib/facter/resolvers/linux/processors_lscpu.rb
